017 refactoring packages names, emit event when doRenewAd to reload the ads in the page, using eventBus as global vue property to coordinate calls between components
