+++ src/popup/backend/adomain/ad/AdStats.js
