+++ src/popup/backend/application/ad/GetAdStatsService.js
