+++ src/popup/backend/application/ad/GetMyFavouriteAdsService.js
