+++ src/popup/backend/framework/ControllerFacadeFactoryBean.js
