+++ src/popup/backend/framework/RepositoryFactoryBean.js
