+++ src/popup/backend/framework/ServiceFactoryBean.js
