+++ src/popup/backend/framework/controller/ad/LoadPageController.js
