+++ src/popup/backend/framework/controller/ad/stats/GetAdStatsController.js
