+++ src/popup/backend/framework/controller/user/DoLogoutController.js
