+++ src/popup/backend/framework/repository/ad/GetAdStatsRepository.js
