+++ src/popup/backend/framework/repository/user/DoLoginWithCookiesRepository.js
