+++ src/popup/event-bus.js
@@ -1,0 +1,2 @@
+import Vue from 'vue';
+export const EventBus = new Vue();
